fixup! update

+++ sdks/java/testing/nexmark/src/main/java/org/apache/beam/sdk/nexmark/NexmarkConfiguration.java
@@ -215,7 +215,7 @@ public class NexmarkConfiguration implements Serializable {
    * Used by Query 13. This specifies number of random keys to generate. Generated events will be
    * assigned to these keys randomly and then goes through a GBK.
    */
-  @JsonProperty public long numKeyBuckets = 20000L;
+  @JsonProperty public int numKeyBuckets = 20000;
 
   /** Replace any properties of this configuration which have been supplied by the command line. */
   public void overrideFromOptions(NexmarkOptions options) {
